Add files via upload
This program makes it possible to obtain confidence regions for estimates of parameters in phase-type distributions. The program is carried out in C by modification of an already existing program (see Marita Olsson, The empht-programme, Relatório técnico, Department of Mathematics, Chalmers University of Technology, and Göteborg University, Sweden, 1998). For more details see chapter 6 in the attached pdf. It is allowed to use this program, provided that authors name, Malte Lindholm Stoltze Rasmussen and link to this page is expressly stated as source.